Added a helper module for common methods for Emotion & Polarity
