+++ lib/sad_panda/emotion.rb
@@ -1,6 +1,10 @@
+require 'sad_panda/helpers'
+
 module SadPanda
   # Emotion determining logic in here
   class Emotion
+    include Helpers
+
     attr_accessor :words, :scores
 
     def initialize(text)
@@ -12,11 +16,8 @@ module SadPanda
 
     # Main method that initiates scoring emotions
     def call
-      remove_stopwords
-
-      stem_words
-
-      score_words(word_frequencies)
+      words = stem_words(remove_stopwords(@words))
+      score_words(get_frequencies_for(words))
 
       scores.key(scores.values.max)
     end
@@ -27,7 +28,7 @@ module SadPanda
     def method_missing(emotion)
       return scores[emotion] || 0 if scores.keys.include? emotion
 
-      raise NoMethodError, 'This method is not defined'
+      raise NoMethodError, "#{emotion} is not defined"
     end
 
     private
@@ -61,33 +62,6 @@ module SadPanda
       end
 
       ambiguous_score
-    end
-
-    # Returns a Hash of frequencies of each uniq word in the text
-    def word_frequencies
-      word_frequencies = {}
-      words.each { |word| word_frequencies[word] = words.count(word) }
-      word_frequencies
-    end
-
-    # Converts all the words to its stem form
-    def stem_words
-      stemmer = Lingua::Stemmer.new(language: 'en')
-      words.map! { |word| stemmer.stem(word) }
-    end
-
-    # Strps the words of stop words
-    def remove_stopwords
-      @words -= SadPanda::Stopwords
-    end
-
-    # Removes all the unwated characters from the text
-    def words_in_text(text)
-      text.downcase!
-      text.gsub!(/((([A-Za-z]{3,9}:(?:\/\/)?)(?:[-;:&=\+\$,\w]+@)?[A-Za-z0-9.-]+|(?:www.|[-;:&=\+\$,\w]+@)[A-Za-z0-9.-]+)((?:\/[\+~%\/.\w-_]*)?\??(?:[-\+=&;%@.\w_]*)#?(?:[\w]*))?)/, '')
-      text.gsub!(/(?=\w*h)(?=\w*t)(?=\w*t)(?=\w*p)\w*/, '')
-      text.gsub!(/\s\s+/, ' ')
-      text.split(' ')
     end
   end
 end
